refactor(test): Moved packages to a feature/domain organization
    What:
    Removed services and db packages and moved children up

    Why:
    This organization makes the sub-systems clearer and
    separates the concerns
