+++ tests/article_classification/test_corruption_adapter.py
